fix: 즉시사용 인수테스트 수정 및 로직 생성
